change datatype to size_t. add function comment

+++ src/bloom.c
@@ -18,13 +18,23 @@
  * Returns:
  *     unsigned integer
  */
-static uint64_t ideal_size(const uint64_t expected, const float accuracy) {
+static uint64_t ideal_size(const size_t expected, const float accuracy) {
 	return -(expected * log(accuracy) / pow(log(2.0), 2));
 }
 
-// TODO: specify which hashing algorithm to use.
-// TODO: add comment/documentation
-bool bloom_init(bloomfilter *bf, const uint64_t expected, const float accuracy) {
+/* bloom_init() -- initialize a bloom filter
+ *
+ * Args:
+ *     bf       - bloomfilter structure
+ *     expected - expected number of elements
+ *     accuracy - margin of acceptable error. ex: 0.01 is "99.99%" accurate
+ *
+ * Returns:
+ *     true on success, false on failure
+ *
+ * TODO: specify which hashing algorithm to use.
+ */
+bool bloom_init(bloomfilter *bf, const size_t expected, const float accuracy) {
 	bf->size        = ideal_size(expected, accuracy);
 	bf->hashcount   = (bf->size / expected) * log(2);
 	bf->bitmap_size = ceil(bf->size / 8);
